Implementation of GPG in order to mask the real password.
It is weell-know that GPG provides an extra and reliable security mechanism whilst communicating sensitive information. The previous method in use for the communication of the password was not robust enoguh. With this modification noone will have access to the real key without the original passphrase (provided by an initialization telegram command).